criacao dos outros verbos da api

+++ GeekShopping/GeekShopping.ProductAPI/Repository/IProductRepository.cs
@@ -5,10 +5,10 @@ namespace GeekShopping.ProductAPI.Repository
     public interface IProductRepository
     {
         Task<IEnumerable<ProductVO>> FindAll();
-        Task<ProductVO> FindById(int id);
+        Task<ProductVO> FindById(long id);
         Task<ProductVO> GetAll();
         Task<ProductVO> Create(ProductVO vo);
         Task<ProductVO> Update(ProductVO vo);
-        Task<bool> Delete(int id);
+        Task<bool> Delete(long id);
     }
 }
